整理几个方法类，使得这些方法能用于非vscode插件环境
